v4l-utils: Split ir-keytable to a specific package
The ir-keytable is useful without the other utilities so this splits
the package out. For flexibility a virtual runtime for the keymap
package is also added so, if necessary, a custom one can be used in a
distro.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>
(cherry picked from commit 3ba7ed2bc83795733a03f19b02fd521c0a00e7e4)
Signed-off-by: Andreas Oberritter <obi@opendreambox.org>