netperf: upgrade to 2.7.0+ with git version
The homepage of netperf has been moved to github. The original webpage
www.netperf.org is redirected to github and tarballs are removed. So use
git version of recipe instead.

netperf 2.7.0 was released in July 2015 which is too old. So upgrade to
latest commit of git repo which is on Nov 29, 2016.

Signed-off-by: Kai Kang <kai.kang@windriver.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>