libtest-warn-perl: add recipe for 0.36
ptest dependency for libcgi-perl

NAME
Test::Warn - Perl extension to test methods for warnings

DESCRIPTION
A good style of Perl programming calls for a lot of diverse regression tests.

This module provides a few convenience methods for testing warning based-code.

If you are not already familiar with the Test::More manpage now would be the time to go take a look.

Signed-off-by: Tim Orling <timothy.t.orling@linux.intel.com>