wavemon: add recipe for version 0.9.5
wavemon is an ncurses-based monitoring application for wireless network
devices on Linux.

We have to provide the path to libnl3 headers since the build system is
not able to find them.

In order to workaround a link issue with pthread library, we have to
add -pthread to CFLAGS in order to add the library after the object
file.

  arm-none-linux-gnueabihf/bin/ld: info_scr.o: undefined reference to symbol 'pthread_mutex_trylock@@GLIBC_2.4'
  [...]/wavemon/0.9.5-r0/recipe-sysroot/lib/libpthread.so.0: error adding symbols: DSO missing from command line

  "We should mention the library on the command line after the object files being compiled" [1]

[1] https://stackoverflow.com/questions/19901934/libpthread-so-0-error-adding-symbols-dso-missing-from-command-line

Signed-off-by: Romain Naour <romain.naour@smile.fr>
Signed-off-by: Khem Raj <raj.khem@gmail.com>